requirements.txt
# ── Core FastAPI stack ────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
starlette>=0.37.0
python-multipart>=0.0.9       # required for Form() parsing

# ── Sessions ──────────────────────────────────────────────────────────────────
itsdangerous>=2.1.2

# ── Templates ─────────────────────────────────────────────────────────────────
jinja2>=3.1.4

# ── Database (async) ──────────────────────────────────────────────────────────
sqlalchemy[asyncio]>=2.0.30
aiosqlite>=0.20.0             # SQLite async driver (local dev / zero config)

# ── PostgreSQL (production) ───────────────────────────────────────────────────
# asyncpg>=0.29.0             # uncomment for PostgreSQL
# psycopg2-binary>=2.9.9      # uncomment for Alembic migrations

# ── Settings & env ────────────────────────────────────────────────────────────
pydantic>=2.7.0
pydantic-settings>=2.2.1
python-dotenv>=1.0.1

# ── ML ────────────────────────────────────────────────────────────────────────
numpy>=1.26.0
scikit-learn>=1.4.0

# ── AI chatbot (optional) ─────────────────────────────────────────────────────
groq>=0.9.0

# ── Production server ─────────────────────────────────────────────────────────
gunicorn>=22.0.0
